requirements.txt
autogen
autogen-agentchat
flask
openai
gunicorn
azure-data-tables>=12.7.0
azure-ai-documentintelligence==1.0.1
